+++ src/utilities/eventCommands/news.js
@@ -1,19 +1,20 @@
 const { EmbedBuilder, ActionRowBuilder, ButtonBuilder, ButtonStyle } = require('discord.js');
 const fetch = require('node-fetch');
-const { handleError } = require('../functions/utilities')
+const { handleError } = require('../functions/utilities');
 
 module.exports = {
     async execute(interaction, auth) {
-        var url = `https://api.mozambiquehe.re/news?auth=${auth}`
+        var url = `https://api.mozambiquehe.re/news?auth=${auth}`;
         fetch(url)
             .then(res => {
                 if (res.status === 200) { return res.json() } else {
-                    handleError(interaction, langOpt, res.status)
+                    handleError(interaction, langOpt, res.status);
                     return Promise.reject('Error occurred');
                 }
             })
             .then(async data => {
-                await interaction.deferUpdate()
+
+                await interaction.deferUpdate();
 
                 if (interaction.message.components[0].components[0].data.disabled === true) {
 
@@ -21,7 +22,7 @@ module.exports = {
                         .setTitle(data[1].title)
                         .setDescription(data[1].short_desc)
                         .setURL(data[1].link)
-                        .setImage(data[1].img)
+                        .setImage(data[1].img);
 
                     const row = new ActionRowBuilder()
                         .addComponents(
@@ -33,16 +34,16 @@ module.exports = {
                                 .setLabel(">")
                                 .setCustomId('2')
                                 .setStyle(ButtonStyle.Success)
-                        )
-                    return interaction.message.edit({ embeds: [embed1], components: [row] })
+                        );
 
+                    return interaction.message.edit({ embeds: [embed1], components: [row] });
                 }
                 if (interaction.customId === '2') {
                     var embed1 = new EmbedBuilder()
                         .setTitle(data[2].title)
                         .setDescription(data[2].short_desc)
                         .setURL(data[2].link)
-                        .setImage(data[2].img)
+                        .setImage(data[2].img);
 
                     const row = new ActionRowBuilder()
                         .addComponents(
@@ -55,16 +56,16 @@ module.exports = {
                                 .setCustomId('2')
                                 .setStyle(ButtonStyle.Danger)
                                 .setDisabled(true)
-                        )
+                        );
 
-                    return interaction.message.edit({ embeds: [embed1], components: [row] })
+                    return interaction.message.edit({ embeds: [embed1], components: [row] });
                 }
                 if (interaction.message.components[0].components[1].data.disabled === true) {
                     var embed1 = new EmbedBuilder()
                         .setTitle(data[1].title)
                         .setDescription(data[1].short_desc)
                         .setURL(data[1].link)
-                        .setImage(data[1].img)
+                        .setImage(data[1].img);
 
                     const row = new ActionRowBuilder()
                         .addComponents(
@@ -76,15 +77,15 @@ module.exports = {
                                 .setLabel(">")
                                 .setCustomId('2')
                                 .setStyle(ButtonStyle.Success)
-                        )
+                        );
 
-                    return interaction.message.edit({ embeds: [embed1], components: [row] })
+                    return interaction.message.edit({ embeds: [embed1], components: [row] });
                 } else {
                     var embed1 = new EmbedBuilder()
                         .setTitle(data[0].title)
                         .setDescription(data[0].short_desc)
                         .setURL(data[0].link)
-                        .setImage(data[0].img)
+                        .setImage(data[0].img);
 
                     const row = new ActionRowBuilder()
                         .addComponents(
@@ -97,9 +98,10 @@ module.exports = {
                                 .setLabel(">")
                                 .setCustomId('2')
                                 .setStyle(ButtonStyle.Success)
-                        )
-                    return interaction.message.edit({ embeds: [embed1], components: [row] })
+                        );
+
+                    return interaction.message.edit({ embeds: [embed1], components: [row] });
                 }
-            }).catch(error => { console.error('Fetch error:', error) })
+            }).catch(error => { console.error('Fetch error:', error) });
     }
 }